requirements.txt
et_xmlfile==2.0.0
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
schedule==1.2.2
six==1.17.0
tzdata==2025.2
